algorithmTemplate_issplit/requirements.txt
django==4.1.7
mysql-connector-python==8.1.0
mysqlclient==2.0.3
pandas==2.1.1
pip==23.3
PyYAML==6.0.1
tqdm==4.65.0
requests==2.28.5